+++ server/src/main/java/edu/byu/cs/tweeter/server/dao/DynamoFeedDAO.java
@@ -1,0 +1,113 @@
+package edu.byu.cs.tweeter.server.dao;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import edu.byu.cs.tweeter.model.domain.Status;
+import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.model.net.request.PagedRequest;
+import edu.byu.cs.tweeter.model.net.response.FeedResponse;
+import edu.byu.cs.tweeter.util.FakeData;
+
+/**
+ * A DAO for accessing 'feed' data from the database.
+ */
+public class DynamoFeedDAO {
+    /**
+     * Gets the count of statuses from the database that the user specified has in their feed. The
+     * current implementation uses generated data and doesn't actually access a database.
+     *
+     * @param user the User whose count of how many statuses is desired.
+     * @return said count.
+     */
+    public Integer getStatusCount(User user) {
+        // TODO: uses the dummy data.  Replace with a real implementation.
+        assert user != null;
+        return getDummyStatuses().size();
+    }
+
+    /**
+     * Gets the statuses from the database that the user specified in the request has in their feed. Uses
+     * information in the request object to limit the number of statuses returned and to return the
+     * next set of statuses after any that were returned in a previous request. The current
+     * implementation returns generated data and doesn't actually access a database.
+     *
+     * @param request contains information about the user whose feed is to be returned and any
+     *                other information required to satisfy the request.
+     * @return the statuses.
+     */
+    public FeedResponse getFeed(PagedRequest<Status> request) {
+        // TODO: Generates dummy data. Replace with a real implementation.
+        assert request.getLimit() > 0;
+        assert request.getUserAlias() != null;
+
+        List<Status> allStatuses = getDummyStatuses();
+        List<Status> responseStatuses = new ArrayList<>(request.getLimit());
+
+        boolean hasMorePages = false;
+
+        if(request.getLimit() > 0) {
+            if (allStatuses != null) {
+                int statusesIndex = getStatusesStartingIndex(request.getLastItem(), allStatuses);
+
+                for(int limitCounter = 0; statusesIndex < allStatuses.size() && limitCounter < request.getLimit(); statusesIndex++, limitCounter++) {
+                    responseStatuses.add(allStatuses.get(statusesIndex));
+                }
+
+                hasMorePages = statusesIndex < allStatuses.size();
+            }
+        }
+
+        return new FeedResponse(responseStatuses, hasMorePages);
+    }
+
+    /**
+     * Determines the index for the first status in the specified 'allStatuses' list that should
+     * be returned in the current request. This will be the index of the next status after the
+     * specified 'lastStatus'.
+     *
+     * @param lastStatus the last status that was returned in the previous
+     *                          request or null if there was no previous request.
+     * @param allStatuses the generated list of statuses from which we are returning paged results.
+     * @return the index of the first status to be returned.
+     */
+    private int getStatusesStartingIndex(Status lastStatus, List<Status> allStatuses) {
+
+        int statusesIndex = 0;
+
+        if(lastStatus != null) {
+            // This is a paged request for something after the first page. Find the first item
+            // we should return
+            for (int i = 0; i < allStatuses.size(); i++) {
+                if(lastStatus.equals(allStatuses.get(i))) {
+                    // We found the index of the last item returned last time. Increment to get
+                    // to the first one we should return
+                    statusesIndex = i + 1;
+                    break;
+                }
+            }
+        }
+
+        return statusesIndex;
+    }
+
+    /**
+     * Returns the list of dummy status data. This is written as a separate method to allow
+     * mocking of the statuses.
+     *
+     * @return the statuses.
+     */
+    List<Status> getDummyStatuses() {
+        return getFakeData().getFakeStatuses();
+    }
+
+    /**
+     * Returns the {@link FakeData} object used to generate dummy statuses.
+     * This is written as a separate method to allow mocking of the {@link FakeData}.
+     *
+     * @return a {@link FakeData} instance.
+     */
+    FakeData getFakeData() {
+        return new FakeData();
+    }
+}
