+++ server/src/main/java/edu/byu/cs/tweeter/server/dao/UserDAO.java
@@ -1,0 +1,9 @@
+package edu.byu.cs.tweeter.server.dao;
+
+import edu.byu.cs.tweeter.model.domain.User;
+
+public interface UserDAO {
+    public void createUser(User user);
+    public User getUser(String username);
+    public void deleteUser(User user);
+}
