+++ server/src/test/java/edu/byu/cs/tweeter/server/service/FollowServiceTest.java
@@ -11,13 +11,13 @@ import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
 import edu.byu.cs.tweeter.model.net.request.PagedRequest;
 import edu.byu.cs.tweeter.model.net.response.FollowingResponse;
-import edu.byu.cs.tweeter.server.dao.FollowDAO;
+import edu.byu.cs.tweeter.server.dao.DynamoFollowDAO;
 
 public class FollowServiceTest {
 
     private PagedRequest<User> request;
     private FollowingResponse expectedResponse;
-    private FollowDAO mockFollowDAO;
+    private DynamoFollowDAO mockDynamoFollowDAO;
     private FollowService followServiceSpy;
 
     @Before
@@ -36,18 +36,18 @@ public class FollowServiceTest {
         // Setup a request object to use in the tests
         request = new PagedRequest<>(authToken, currentUser.getAlias(), 3, null);
 
-        // Setup a mock FollowDAO that will return known responses
+        // Setup a mock DynamoFollowDAO that will return known responses
         expectedResponse = new FollowingResponse(Arrays.asList(resultUser1, resultUser2, resultUser3), false);
-        mockFollowDAO = Mockito.mock(FollowDAO.class);
-        Mockito.when(mockFollowDAO.getFollowees(request)).thenReturn(expectedResponse);
+        mockDynamoFollowDAO = Mockito.mock(DynamoFollowDAO.class);
+        Mockito.when(mockDynamoFollowDAO.getFollowees(request)).thenReturn(expectedResponse);
 
         followServiceSpy = Mockito.spy(FollowService.class);
-        Mockito.when(followServiceSpy.getFollowingDAO()).thenReturn(mockFollowDAO);
+        Mockito.when(followServiceSpy.getFollowingDAO()).thenReturn(mockDynamoFollowDAO);
     }
 
     /**
      * Verify that the {@link FollowService#getFollowees(PagedRequest<User>)}
-     * method returns the same result as the {@link FollowDAO} class.
+     * method returns the same result as the {@link DynamoFollowDAO} class.
      */
     @Test
     public void testGetFollowees_validRequest_correctResponse() {
